Adding new validation, refactor of validation
I previously added a validation and couldn't get it to trigger. I
realized that since we're doing 2 passes on validation, some of the
validation I was adding wasn't getting triggered.

Ideally, we parse and validate the document in 1 pass. I've moved all
validation to the parser. This will allow us to move all the defensive
code to the parser as well.