Move plugin and schema.json out of internal repo
This moves the schema.json out into the light and puts it with the rest
of the parsing and validating code where it belongs. Since they change
together and we should publish them together.

I removed checkfile related things from the plugin, we can decide what
to do about that. I also moved the JSON Schema checks to inside this
library instead of being in the plugin. We should add more unit testing
on this library to exercise and improve the JSON schema validation.

We can add some light XTP plugin tests to make sure that the plugin is
still working as behaving, but not going so far as to create a lot of
individual test cases that couple the wasm level tests to the
implementation details of things like features and error messages.